Fix flaky test and cleanup a little bit of tech debt
