+++ runtime/defaults/defaults.go
@@ -13,6 +13,7 @@ const (
 	defaultRPCTimeout = 30000
 )
 
+// TODO: some of these defaults are used for production while others are used for testing. Need to separate them.
 var (
 	DefaultChains              = []string{"0001"}
 	DefaultServiceURL          = ""
@@ -22,6 +23,8 @@ var (
 	DefaultMaxRelaysString     = converters.BigIntToString(DefaultMaxRelays)
 	DefaultAccountAmount       = big.NewInt(100000000000000)
 	DefaultAccountAmountString = converters.BigIntToString(DefaultAccountAmount)
+	DefaultSendAmount          = big.NewInt(10000)
+	DefaultSendAmountString    = converters.BigIntToString(DefaultSendAmount)
 	DefaultPauseHeight         = int64(-1)
 	DefaultUnstakingHeight     = int64(-1)
 	DefaultChainID             = "testnet"
